server-project/requirements.txt
alembic==1.11.1
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mongoengine==0.27.0
packaging==23.1
pycparser==2.21
PyJWT==2.7.0
pymongo==4.3.3
PyMySQL==1.0.3
python-dotenv==1.0.0
python-http-client==3.3.7
sendgrid==6.10.0
six==1.16.0
SQLAlchemy==2.0.15
starkbank-ecdsa==2.2.0
typing_extensions==4.6.2
Werkzeug==2.3.4
WTForms==3.0.1
